Updated the testsuite to the new structure
